First step to simplify grid GUI to use components rather than exact coords
Previously, exact coordinates were used to draw the grid and cell numbers, which made it a tad complicated when determining things like where the mouse clicked on due to having to calculate where on the grid the click would be located in. The swap to using components to represent cells should help simplify all of this.