Fix performance bug and add GUI enhancements
Performance bug was me not realising that setFont() called repaint, so an infinite loop was occurring...
Added different colors to the grid to differentiate unchangeable cells.
Added the border for unchangeable cells so that it's clearer which unchangeable cell is selected.

I also added some containers to store related and adjacent cellLabels to make some stuff easier.
This was originally done in an attempt to make some calculatations quicker while I was trying to debug the performance bug.
I've mainly left this in because it makes it simple to check all related cells to a single selected cell.
Not sure if I'll keep it in or add to it, it feels kind of pointless to keep it for a single method.